third_party/WebKit/LayoutTests/fast/css/preferred-stylesheet-reversed-order.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<div id="t1">This text should be green</div>
<script>
description("Preferred stylesheet where insertion order is tree order");

function createStyleElement(text, title) {
    var elm = document.createElement("style");
    elm.setAttribute("title", title);
    elm.appendChild(document.createTextNode(text));
    return elm;
}

document.head.insertBefore(createStyleElement("#t1 {color:green}", "preferred"), document.head.firstChild);
document.head.insertBefore(createStyleElement("#t1 {color:red}", "notpreferred"), document.head.firstChild);

shouldBeEqualToString("getComputedStyle(t1).color", "rgb(0, 128, 0)");
shouldBeEqualToString("document.preferredStylesheetSet", "preferred");
shouldBeEqualToString("document.selectedStylesheetSet", "preferred");
</script>
